SC requests moved to its own file. Testing getTracks in Staging.
